Update PR comments
